refactor boot runtime assembly
