--- yak-ops-boot/src/main/java/io/baize/flow/api/config/SchedulingConfig.java
@@ -1,9 +1,0 @@
-package io.baize.flow.api.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@Configuration
-@EnableScheduling
-public class SchedulingConfig {
-}
