+++ yak-ops-boot/src/main/java/io/baize/flow/boot/BaizeFlowApplication.java
@@ -1,19 +1,25 @@
-package io.baize.flow.api;
+package io.baize.flow.boot;
 
-import jakarta.annotation.Resource;
+import io.baize.flow.dao.DaoConfiguration;
 import org.apache.seatunnel.plugin.datasource.api.plugin.DataSourceProcessorProvider;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Import;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
-
-@SpringBootApplication(scanBasePackages = {"org.apache.seatunnel.*"})
+@SpringBootApplication
+@ComponentScan(basePackages = {
+        "io.baize.flow.api",
+        "io.baize.flow.application",
+        "io.baize.flow.infrastructure",
+        "io.baize.flow.dao",
+        "io.baize.flow.engine",
+        "org.apache.seatunnel.plugin.datasource.api"
+})
+@Import(DaoConfiguration.class)
 @EnableConfigurationProperties
-@EnableScheduling
 @EnableAsync(proxyTargetClass = true)
 public class BaizeFlowApplication {
 
@@ -27,4 +33,3 @@ public class BaizeFlowApplication {
         }
     }
 }
-
