+++ yak-ops-boot/src/main/java/io/baize/flow/boot/config/SchedulingConfig.java
@@ -1,0 +1,11 @@
+package io.baize.flow.boot.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+
+@Configuration
+@EnableScheduling
+@ConditionalOnProperty(prefix = "yak-ops.scheduler", name = "enabled", havingValue = "true", matchIfMissing = true)
+public class SchedulingConfig {
+}
