+++ yak-ops-boot/src/main/java/io/baize/flow/boot/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package io.baize.flow.api.config;
+package io.baize.flow.boot.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
