+++ yak-ops-dao/pom.xml
@@ -47,10 +47,6 @@
             <artifactId>yak-ops-datasource-support</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.baize.flow</groupId>
-            <artifactId>yak-ops-dao-plugin-all</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.zaxxer</groupId>
             <artifactId>HikariCP</artifactId>
         </dependency>
